+++ marketplace/ui/src/components/MarketPlace/index.js
@@ -16,6 +16,20 @@ import { SEO } from "../../helpers/seoConstant";
 import { BANNER } from "../../helpers/constants";
 import { bannerArrow } from "../../images/SVGComponents";
 
+// ----------------------------------------------------------
+
+import { Swiper, SwiperSlide } from 'swiper/react';
+
+// Import Swiper styles
+import 'swiper/css';
+import 'swiper/css/effect-fade';
+import 'swiper/css/navigation';
+import 'swiper/css/pagination';
+import 'swiper/css/autoplay';
+
+// import required modules
+import { EffectFade, Navigation, Pagination, Autoplay } from 'swiper/modules';
+
 
 const MarketPlace = ({ user, isAuthenticated }) => {
   const limit = 10, offset = 0;
@@ -65,22 +79,22 @@ const MarketPlace = ({ user, isAuthenticated }) => {
   }, [dispatch, limit, offset, debouncedSearchTerm]);
 
   const linkUrl = window.location.href;
-  const navRoute = routes.MarketplaceCategoryProductList.url.replace(':category', 'All');
+  // const navRoute = routes.MarketplaceCategoryProductList.url.replace(':category', 'All');
 
-  const ButtonElement = ({ desktopText, mobileText }) => <div className="relative flex top-[156px] sm:top-[250px] xl:top-[65%] 3xl:top-[70%] left-[4%] sm:left-[7.5%] md:left-[7%] md:top-60 z-50">
+  const ButtonElement = ({ desktopText, mobileText, url }) => <div className="relative flex top-[156px] sm:top-[250px] xl:top-[65%] 3xl:top-[70%] left-[4%] sm:left-[7.5%] md:left-[7%] md:top-60 z-50">
     <Button
       id="viewMore"
       onClick={() => {
-        navigate(navRoute);
+        navigate(url);
         sessionStorage.setItem('scrollPosition', 0);
       }}
       className="gradient-button border-0 h-auto md:h-11 border-primary bg-white text-primary hover:text-white"
     >
       <div className="flex items-center">
-        <div className="hidden sm:block font-semibold text-lg">
+        <div className="hidden sm:block font-semibold text-lg banner-btn-text">
           {desktopText}
         </div>
-        <div className="sm:hidden font-semibold text-base">
+        <div className="sm:hidden font-semibold text-base banner-btn-text">
           {mobileText}
         </div>
         <span className="ml-1">{bannerArrow}</span>
@@ -89,60 +103,74 @@ const MarketPlace = ({ user, isAuthenticated }) => {
   </div>
 
   const CarouselElement = ({ scrollT }) =>
-    <Carousel autoPlay centerSlidePercentage={95} showArrows={false} infiniteLoop showStatus={false} swipeable emulateTouch autoFocus centerMode
-      preventMovementUntilSwipeScrollTolerance={true} swipeScrollTolerance={scrollT}
+    <Swiper
+      spaceBetween={30}
+      effect={'fade'}
+      navigation={false}
+      centeredSlides={true}
+      pagination={{
+        clickable: true,
+      }}
+      autoplay={{
+        delay: 4000,
+        disableOnInteraction: false,
+      }}
+      modules={[Autoplay, EffectFade, Navigation, Pagination]}
+      className="mySwiper"
     >
       {BANNER.map((item, index) => (
-        <div key={index} className="no-select relative p-2 h-[222px] sm:h-[380px] 3xl:h-[480px] mx-1 md:mx-2 md:mt-6 lg:mx-3">
-          <ButtonElement desktopText={item.desktopText} mobileText={item.mobileText} />
+        <SwiperSlide>
+          <div key={index} className="no-select relative p-2 h-[222px] sm:h-[380px] 3xl:h-[480px] mx-1 md:mx-2 md:mt-6 lg:mx-3">
+            <ButtonElement desktopText={item.desktopText} mobileText={item.mobileText} url={item.link} />
             {item.text}
-          <div className="sm:hidden">
-            <img
-              alt={item.alt}
-              title={item.title}
-              className="no-select absolute inset-0 z-10 h-[222px] md:w-[90%] rounded-md md:rounded-[14px] drop-shadow-md"
-              height={380}
-              width="100%"
-              src={item.mobileImg}
-              preview={false}
-            />
+            <div className="sm:hidden">
+              <img
+                alt={item.alt}
+                title={item.title}
+                className="no-select absolute inset-0 z-10 h-[222px] w-[96%] rounded-md md:rounded-[14px] drop-shadow-md mx-auto"
+                height={380}
+                width="100%"
+                src={item.mobileImg}
+                preview={false}
+              />
+            </div>
+            <div className="hidden sm:block md:hidden">
+              <img
+                alt={item.alt}
+                title={item.title}
+                className="no-select absolute inset-0 z-10 h-[380px] w-[96%] rounded-md md:rounded-[14px] drop-shadow-md mx-auto"
+                height={380}
+                width="100%"
+                src={item.tabletImg}
+                preview={false}
+              />
+            </div>
+            <div className="hidden md:block lg:hidden">
+              <img
+                alt={item.alt}
+                title={item.title}
+                className="no-select absolute inset-0 z-10 md:h-[330px] w-[98%] rounded-md md:rounded-[14px] drop-shadow-md mx-auto"
+                height={380}
+                width="100%"
+                src={item.laptopImg}
+                preview={false}
+              />
+            </div>
+            <div className="hidden lg:block">
+              <img
+                alt={item.alt}
+                title={item.title}
+                className="no-select absolute inset-0 z-10 lg:h-[330px] 3xl:h-[480px] w-[98%] rounded-md md:rounded-[14px] drop-shadow-md mx-auto"
+                height={380}
+                width="100%"
+                src={item.desktopImg}
+                preview={false}
+              />
+            </div>
           </div>
-          <div className="hidden sm:block md:hidden">
-            <img
-              alt={item.alt}
-              title={item.title}
-              className="no-select absolute inset-0 z-10 h-[380px] md:w-[90%] rounded-md md:rounded-[14px] drop-shadow-md"
-              height={380}
-              width="100%"
-              src={item.tabletImg}
-              preview={false}
-            />
-          </div>
-          <div className="hidden md:block lg:hidden">
-            <img
-              alt={item.alt}
-              title={item.title}
-              className="no-select absolute inset-0 z-10 md:h-[330px] md:w-[90%] rounded-md md:rounded-[14px] drop-shadow-md"
-              height={380}
-              width="100%"
-              src={item.laptopImg}
-              preview={false}
-            />
-          </div>
-          <div className="hidden lg:block">
-            <img
-              alt={item.alt}
-              title={item.title}
-              className="no-select absolute inset-0 z-10 lg:h-[330px] 3xl:h-[480px] md:w-[90%] rounded-md md:rounded-[14px] drop-shadow-md"
-              height={380}
-              width="100%"
-              src={item.desktopImg}
-              preview={false}
-            />
-          </div>
-        </div>
+        </SwiperSlide>
       ))}
-    </Carousel>
+    </Swiper>
 
   return (
     <>
@@ -151,12 +179,7 @@ const MarketPlace = ({ user, isAuthenticated }) => {
         description={SEO.DESCRIPTION_META}
         link={linkUrl} />
       <Fade triggerOnce>
-        <div className="hidden md:block">
           <CarouselElement scrollT={130} />
-        </div>
-        <div className="md:hidden">
-          <CarouselElement scrollT={50} />
-        </div>
       </Fade>
       {iscategorysLoading ? (
         <div className="h-96 flex justify-center items-center">